+++ tests/integrated/test-get.R
@@ -23,7 +23,7 @@ test_that("get_groups works", {
   groups_list <- get_groups(test_conn_1)
   expect_is(groups_list, c("tbl_df", "tbl", "data.frame"))
 
-  expect_equal(vctrs::vec_ptype(groups_list), vctrs::vec_ptype(connectapi_ptypes$groups))
+  expect_ptype_equal(groups_list, connectapi_ptypes$groups)
 })
 
 test_that("get_content works", {
@@ -31,30 +31,33 @@ test_that("get_content works", {
   content_list <- get_content(test_conn_1)
   expect_is(content_list, c("tbl_df", "tbl", "data.frame"))
 
-  # https://github.com/r-lib/testthat/issues/985
-  skip("currently segfaults")
-  expect_equal(vctrs::vec_ptype(content_list), vctrs::vec_ptype(connectapi_ptypes$content))
+  # various attributes have been added over the years, so exact match
+  # doesn't work against all versions of Connect
+  expect_ptype_equal(content_list, connectapi_ptypes$content, exact = FALSE)
 })
 
 test_that("get_usage_shiny works", {
   shiny_usage <- get_usage_shiny(test_conn_1)
   expect_is(shiny_usage, c("tbl_df", "tbl", "data.frame"))
 
-  expect_equal(vctrs::vec_ptype(shiny_usage), vctrs::vec_ptype(connectapi_ptypes$usage_shiny))
+  expect_ptype_equal(shiny_usage, connectapi_ptypes$usage_shiny)
 })
 
 test_that("get_usage_static works", {
   content_visits <- get_usage_static(test_conn_1)
   expect_is(content_visits, c("tbl_df", "tbl", "data.frame"))
 
-  expect_equal(vctrs::vec_ptype(content_visits), vctrs::vec_ptype(connectapi_ptypes$usage_static))
+  # path was added to usage_static in 2024
+  expect_ptype_equal(content_visits, connectapi_ptypes$usage_static, exact = FALSE)
 })
 
 test_that("get_audit_logs works", {
   audit_list <- get_audit_logs(test_conn_1)
   expect_is(audit_list, c("tbl_df", "tbl", "data.frame"))
 
-  expect_equal(vctrs::vec_ptype(audit_list), vctrs::vec_ptype(connectapi_ptypes$audit_logs))
+  # This is different on older versions, not sure it's worth worrying about how
+  skip_if_connect_older_than(test_conn_1, "2022.09.0")
+  expect_ptype_equal(audit_list, connectapi_ptypes$audit_logs)
 })
 
 test_that("get_procs works", {
@@ -64,7 +67,7 @@ test_that("get_procs works", {
   # TODO: This is not a great test, since no processes are running
   # we could always start a content restoration...
   expect_is(proc_data, "tbl_df")
-  expect_equal(vctrs::vec_ptype(proc_data), vctrs::vec_ptype(connectapi_ptypes$procs))
+  expect_ptype_equal(proc_data, connectapi_ptypes$procs)
 })
 
 # experimental --------------------------------------------